refactor: change token and message counts from u64 to usize for consistency and improved memory usage

+++ crates/forge_app/src/app.rs
@@ -151,7 +151,7 @@ impl<S: Services> ForgeApp<S> {
         };
 
         // Calculate original metrics
-        let original_messages = context.messages.len() as u64;
+        let original_messages = context.messages.len();
         let original_tokens = context.token_count();
 
         // Find the main agent (first agent in the conversation)
@@ -175,7 +175,7 @@ impl<S: Services> ForgeApp<S> {
                 .await?;
 
             // Calculate compacted metrics
-            let compacted_messages = compacted_context.messages.len() as u64;
+            let compacted_messages = compacted_context.messages.len();
             let compacted_tokens = compacted_context.token_count();
 
             // Update the conversation with the compacted context
